Add files via upload
Air travel has become an essential mode of transportation for millions of people worldwide, and with the increasing number of passengers, the quality of service provided by airlines has gained significant importance. Passenger satisfaction plays a critical role in the competitive airline industry, influencing not only customer loyalty but also the overall reputation and success of an airline. Understanding what factors contribute to a positive or negative travel experience is essential for airlines aiming to enhance service standards and customer retention.
This project, titled "Airline Passenger Satisfaction and Analysis", is based on a comprehensive dataset that captures various factors affecting the travel experience of airline passengers. It includes demographic information (such as age and gender), travel details (customer type, class, type of travel, flight distance), as well as service quality ratings across multiple touchpoints of a journey.
This dataset provides a detailed snapshot of various aspects influencing airline passenger satisfaction. It contains data on individual passengers including demographic details (like gender and age), type of travel, class of service, and flight-related information (distance, delays, etc.). In addition, it captures customer feedback on multiple service areas such as:
•	Seat comfort and legroom service
•	Cleanliness and food quality
•	In-flight entertainment and Wi-Fi service
•	Baggage handling and customer service
The most valuable aspect of this dataset is the final satisfaction label, which helps in distinguishing between satisfied and dissatisfied passengers based on their overall experience.